Made a separate grammar and illegal value file
